requirements.txt
faiss-cpu
huggingface-hub
logging
numpy
ollama
string
